test(cli): split strategy-profile-diff out of optimize execution tests
test_commands_execution.py (456 lines, allowlisted) mixed two concerns:
the optimize subcommands (run/list/view/compare/export/apply) and
TestStrategyProfileDiffCommand, which exercises
cli.commands.strategy_profile — a different module that was also misplaced
under optimize/.

Extract that class to cli/commands/test_strategy_profile_diff.py (mirrors
the source location). Original drops to 312 lines; remove its now-stale
size-allowlist entry. 17 tests unchanged and passing (12 + 5).
